Move `[data-turbo="false"]` guard up to observers
Since the `FormSubmitObserver` and `LinkClickObserver` classes are
very interested in the page's HTML, re-use the `elementIsNavigable`
checks from the `Session` by importing them into each observer, instead
of the `Session` itself.

This enables the `Session` to act with more confidence that any
interaction on the page is to be Turbo-enabled.