Render 4xx-5xx responses within frame
The use case
---

Consider submitting a `<form data-turbo-frame="_top">` from within a
`<turbo-frame>`:

```html
<dialog open>
  <turbo-frame id="dialog">
    <form data-turbo-target="_top" method="POST" action="/posts">
      <!-- ... -->
    </form>
  </turbo-frame>
</dialog>
```

In the case of success, the response will be a [303 redirecting to
another URL][303].

In the case of a failed submission, the response will be a [422 with an
HTML body][]:

```ruby
class PostsController < ApplicationController
  def create
    @post = Post.create!(post_params)

    redirect_to post_url(@post)
  rescue ActiveRecord::RecordNotSaved
    render inline: <<~HTML, status: :unprocessable_entity
      <turbo-frame id="dialog">
        <form data-turbo-target="_top" method="POST" action="/posts">
          <!-- ... -->
        </form>
      </turbo-frame>
    HTML
  end
end
```

[303]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/303
[422]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/422

The current behavior
---

Since the `<form>` element declares its target as `_top`, a successful
submission resulting in the [303][] response, the entire page redirect.

Similarly, in the case of an invalid submission, _the entire page's
HTML_ will be replaced by the response, in spite of the submission
occurring within a `<turbo-frame>` element.

The desired outcome
---

A valid submission behaves as expected and desired: redirect the `_top`
(entire page) to the URL.

In the case of a submission without Turbo, a failed submission would
result in a full re-render of the page, keeping the `<dialog>`
containing the form open. Any information or context in the "background"
would be re-rendered, but would exist.

With the introduction of Turbo Frames to the situation as a progressive
enhancement, a developer might expect the server's response to be
limited to only replacing the contents of the `<turbo-frame>` after a
failure.

A potential work-around
---

As a workaround, the server could respond with a `<turbo-stream>` to
replace the inner contents of the `<dialog>` element's `<turbo-frame>`.

Proposed change
---

Currently, the implementation treats submissions from within a
`<turbo-frame>` that target `_top` differently from elements without a
`[data-turbo-frame]` attribute, or one that refers to another element on
the page.

To account for the proposed change in behavior, _all_ frame submissions
would use the same machinery, and the different treatment would be
deferred to _after_ the submission is determined to be a success.