Introduce `turbo:frame-missing` event
Closes [hotwired/turbo#432][]
Follow-up to [hotwired/turbo#94][]
Follow-up to [hotwired/turbo#31][]

When a response from _within_ a frame is missing a matching frame, fire
the `turbo:frame-missing` event.

There is an existing [contract][] that dictates a request from within a
frame stays within a frame.

However, if an application is interested in reacting to a response
without a frame, dispatch a `turbo:frame-missing` event. The event's
`target` is the `FrameElement`, and the `detail` contains the
`fetchResponse:` key. To transform the `FetchResponse` into a `Visit`,
clients can extract:

* `redirected: boolean`
* `statusCode: number`
* `responseHTML: Promise<string>`

Event listeners for `turbo:frame-missing` can forward the `detail`
directly to a `Turbo.visit` call:

```js
addEventListener("turbo:frame-missing", async ({ target, detail: { fetchResponse } }) => {
  // the details of `shouldRedirectOnMissingFrame(element: FrameElement)`
  // are up to the application to decide
  if (shouldRedirectOnMissingFrame(target)) {
    const { location, redirected, statusCode, responseHTML } = fetchResponse
    const response = { redirected, statusCode, responseHTML: await responseHTML }

    Turbo.visit(location, { response })
  }
})
```

[contract]: https://github.com/hotwired/turbo/issues/94#issuecomment-756968205
[hotwired/turbo#432]: https://github.com/hotwired/turbo/issues/432
[hotwired/turbo#94]: https://github.com/hotwired/turbo/issues/94
[hotwired/turbo#31]: https://github.com/hotwired/turbo/issues/31