Functional Tests: Replace `chai` with Playwright Assertions
Utilize Playwright's built-in [assertions][] so that functional test
assertions are consistent. The `chai`-based assertions provide similar
coverage, but not much is gained for the suite by having two competing
styles of assertion.

The majority of this diff replaces `assert.equal` with
`expect(...).toEqual()`. In some cases, text comparison is replaced with
Playwright's [toHaveText][], since that assertion bakes in
synchronization and waiting support, and has the ability to compare
contents in a way that ignores spacing differences.

There are other patterns throughout the test suite that could be
improved by adhering more closely to Playwright idioms, but that would
can be saved for another time.

Related
---

A reduction in functional test suite execution time was an unexpected
benefit. Since the suite relies less on manual timing interventions like
`nextBody` and `nextBeat`, it relies more on Playwright's built-in
synchronization and timing mechanisms.

```sh
yarn test:browser --project=chrome

 # ...

  352 passed (38s)
✨  Done in 38.17s.
```

```sh
yarn test:browser --project=chrome

 # ...

  349 passed (25s)
✨  Done in 24.89s.
```

[assertions]: https://playwright.dev/docs/test-assertions
[toHaveText]: https://playwright.dev/docs/api/class-locatorassertions#locator-assertions-to-have-text